refactor(transition): remove transition state 'STATE_INIT'
Previously, when the component was just mounted, the Transition component would be in the
‘STATE_INIT’ state. At this time, the children would not be displayed. As the state of the component
changes, the subcomponents are displayed. So there is always a short period of time that will make
children not show. Now that ‘STATE_INIT’ has been removed, whether or not the sub-component is
displayed depends only on the prop ‘in’, making the component state more in line with expectations.