refactor(popper): use 'padding' to complete 'offset'
In the previous implementation, there was a gap of offset size between the popper and the reference
element. When the user's mouse moves too slowly during the movement from the reference element to
the popper, the popper closes. Now padding is used to fill this gap and improve the user experience.